pyproject.toml
[tool.poetry]
name = "budget-api"
version = "0.1.0"
description = ""
authors = ["MateDawid <mateusiakdawid@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.10"
djangorestframework = "^3.15.2"
dynaconf = {extras = ["yaml"], version = "^3.2.0"}
django-extensions = "^3.2.3"
drf-yasg = "^1.21.7"
psycopg2-binary = "2.9.6"
drf-nested-routers = "^0.93.5"
django-filter = "^24.2"
django-debug-toolbar = "^4.4.2"
pip = "^24.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.8.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
safety = "^3.2.7"
bandit = "^1.7.9"

[tool.pytest.ini_options]
minversion = "7.4.0"
testpaths = [
    "tests"
]
DJANGO_SETTINGS_MODULE = "app_config.settings"
pythonpath  = [
    "src",
    "tests"
]

[tool.black]
line-length = 120
extend-exclude = '''
(
  ^/.git
  |./venv/*
  |./__pycache__/*
  |./.pytest_cache/*
  |./htmlcov/*
  |./migrations/*
  |./docs/*
  |./docs/*
  |./manage.py
  |./docker/*
  |./.github/*"
)
'''

[tool.isort]
profile = 'black'
line_length = 120
skip_glob = [
    ".git",
    "venv/*",
    "*/__pycache__/*",
    "*/.pytest_cache/*",
    "htmlcov/*",
    "*/migrations/*",
    "*/docs/*",
    "*/docs/*",
    "manage.py",
    "*/docker/*",
    "*/.github/*"
]

[tool.bandit]
exclude_dirs = ["tests", "venv", "__pycache__", ".pytest_cache", "htmlcov", "docker", ".github"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
